ref: moved DesktopSideBar component into SideBar/variants
Relocated the DesktopSideBar component into the /SideBar/variants/ directory. The reason for this change is that with the existence of the /SideBar directory the SideBar-Variants directory has lost its meaning.
